added orbital camera demo to visualize cosine distribution, mouse input & various refactorings

+++ src/Kea/core.nim
@@ -4,6 +4,8 @@ import
   colors,
   mesh,
   input,
+  primitives,
+  transform,
   pbr,
   ltc,
   nimgl/[glfw, opengl]
@@ -17,7 +19,8 @@ type
     delta*: float32
     fps*: float32
     time*: float32
-    input*: Input
+    keyboard*: Keyboard
+    mouse*: Mouse
 
   KeaObj = object
     width: Natural
@@ -27,79 +30,67 @@ type
     frameWidth: Natural
     frameHeight: Natural
 
+    window: GLFWWindow
+
+    mouse: Mouse
+    keyboard: Keyboard
+
     passes: seq[RenderPass]
 
     pbr*: PBRRenderer
 
     camera*: Camera
 
-    window: GLFWWindow
-
     storage: MeshStorage
 
     ltc: LtcTextures
-
-    input: Input
 
   Kea* = ref KeaObj
 
 proc `=destroy`(kea: var KeaObj) =
   {.cast(raises: []).}:
-    kea.passes.setLen(0) # ensure renderers deallocate before context ends
+    kea.passes = @[]
 
     kea.pbr = nil
 
-    kea.storage.destroy()
+    if kea.storage != nil:
+      kea.storage.destroy()
+      kea.storage = nil
+
+    kea.ltc.destroy()
+
+    kea.ltc = default(LtcTextures)
+
+    kea.title = ""
 
     if kea.window != nil:
       kea.window.destroyWindow()
+      kea.window = nil
 
     glfwTerminate()
-
-proc errorCallback(error: int32, description: cstring) {.cdecl.} =
-  echo "GLFW error ", error, ": ", description
-
-proc windowSizeCallback(
-    window: GLFWWindow,
-    width: int32,
-    height: int32,
-) {.cdecl.} =
-  let kea = cast[Kea](window.getWindowUserPointer())
-
-  kea.width = Natural(width)
-  kea.height = Natural(height)
-
-proc framebufferSizeCallback(
-    window: GLFWWindow,
-    width: int32,
-    height: int32,
-) {.cdecl.} =
-  let kea = cast[Kea](window.getWindowUserPointer())
-
-  kea.frameWidth = Natural(width)
-  kea.frameHeight = Natural(height)
-
-  glViewport(0, 0, width, height)
 
 proc initKea*(
     width: Natural,
     height: Natural,
     title: string,
-    vertexCapacity: Natural = DEFAULT_VERTEX_CAPACITY,
-    indexCapacity: Natural = DEFAULT_INDEX_CAPACITY,
+    vertexCapacity: Natural = DefaultVertexCapacity,
+    indexCapacity: Natural = DefaultIndexCapacity,
     resizable = false,
 ): Kea =
   when not defined(release):
-    discard glfwSetErrorCallback(errorCallback)
+    discard glfwSetErrorCallback(
+      proc(error: int32, description: cstring) {.cdecl.} =
+        echo "GLFW error ", error, ": ", description
+    )
 
   doAssert glfwInit(), "Failed to initialize GLFW"
 
   if not resizable:
-    glfwWindowHint(GLFWResizable, GLFW_FALSE)
+    glfwWindowHint(GLFWResizable, GLFWFalse)
 
   glfwWindowHint(GLFWContextVersionMajor, 3)
   glfwWindowHint(GLFWContextVersionMinor, 3)
-  glfwWindowHint(GLFWOpenglProfile, GLFW_OPENGL_CORE_PROFILE)
+  glfwWindowHint(GLFWOpenglProfile, GLFWOpenglCoreProfile)
   glfwWindowHint(GLFWSamples, 8)
 
   let window = glfwCreateWindow(
@@ -153,10 +144,45 @@ proc initKea*(
     passes: @[pbrPass]
   )
 
+  var cursorX, cursorY: float64
+  window.getCursorPos(addr cursorX, addr cursorY)
+
+  result.mouse.position = [cursorX.float32, cursorY.float32]
+
   window.setWindowUserPointer(cast[pointer](result))
 
-  discard window.setWindowSizeCallback(windowSizeCallback)
-  discard window.setFramebufferSizeCallback(framebufferSizeCallback)
+  discard window.setWindowSizeCallback(
+    proc(
+      window: GLFWWindow,
+      width, height: int32,
+    ) {.cdecl.} =
+      let kea = cast[Kea](window.getWindowUserPointer())
+
+      kea.width = Natural(width)
+      kea.height = Natural(height)
+  )
+
+  discard window.setFramebufferSizeCallback(
+    proc(
+      window: GLFWWindow,
+      width, height: int32,
+    ) {.cdecl.} =
+      let kea = cast[Kea](window.getWindowUserPointer())
+
+      kea.frameWidth = Natural(width)
+      kea.frameHeight = Natural(height)
+
+      glViewport(0, 0, width, height)
+  )
+
+  discard window.setScrollCallback(
+    proc(
+      window: GLFWWindow,
+      xOffset, yOffset: float64,
+    ) {.cdecl.} =
+      let kea = cast[Kea](window.getWindowUserPointer())
+      kea.mouse.scroll = [xOffset.float32, yOffset.float32]
+  )
 
 proc newMesh*(
     kea: Kea,
@@ -165,12 +191,109 @@ proc newMesh*(
 ): Mesh =
   mesh.new(kea.storage, vertices, indices)
 
-proc newRenderer*[T](
+proc newRenderer*[T: tuple](
   kea: Kea,
   frag: string,
   vert = DefaultVert,
 ): Renderer[T] =
-  renderer.new(vert = vert, frag = frag, kea.storage)
+  renderer.new(kea.storage, vert = vert, frag = frag)
+
+proc add*[T](
+  kea: Kea,
+  mesh: Mesh,
+  frag: string,
+  vert = DefaultVert,
+  material: T = (),
+  transform = Identity
+): Drawable[T] =
+  let renderer = renderer.new[T](
+    kea.storage,
+    frag = frag,
+    vert = vert
+  )
+
+  kea.passes.add RenderPass(
+    render: proc(ctx: RenderContext) =
+      renderer.render(ctx)
+  )
+
+  renderer.add(mesh, material, transform)
+
+proc add*[T](
+  kea: Kea,
+  mesh: Mesh,
+  frag: string,
+  vert = DefaultVert,
+  material: T = (),
+  x: float32 = 0.0,
+  y: float32 = 0.0,
+  z: float32 = 0.0,
+  yaw: float32 = 0.0,
+  pitch: float32 = 0.0,
+  roll: float32 = 0.0,
+  scale: float32 = 1.0
+): Drawable[T] = 
+  kea.add(
+    mesh = mesh,
+    frag = frag,
+    vert = vert,
+    material = material,
+    transform = transform.new(
+      x = x,
+      y = y,
+      z = z,
+      pitch = pitch,
+      yaw = yaw,
+      roll = roll,
+      scale = scale
+    )
+  )
+
+proc add*[T](
+  kea: Kea,
+  primitive: Primitive,
+  frag: string,
+  vert = DefaultVert,
+  material: T = (),
+  transform = Identity
+): Drawable[T] =
+  kea.add(
+    mesh = primitive.mesh(kea.storage), 
+    frag = frag, 
+    vert = vert, 
+    material = material,
+    transform = transform
+  )
+
+proc add*[T](
+  kea: Kea,
+  primitive: Primitive,
+  frag: string,
+  vert = DefaultVert,
+  material: T = (),
+  x: float32 = 0.0,
+  y: float32 = 0.0,
+  z: float32 = 0.0,
+  yaw: float32 = 0.0,
+  pitch: float32 = 0.0,
+  roll: float32 = 0.0,
+  scale: float32 = 1.0
+): Drawable[T] =
+  kea.add(
+    mesh = primitive.mesh(kea.storage), 
+    frag = frag, 
+    vert = vert, 
+    material = material,
+    transform = transform.new(
+      x = x,
+      y = y,
+      z = z,
+      pitch = pitch,
+      yaw = yaw,
+      roll = roll,
+      scale = scale
+    )
+  )
 
 proc render*(kea: Kea, clear: Color = [0.1, 0.1, 0.1, 1.0]) =
   if kea.frameWidth == 0 or kea.frameHeight == 0:
@@ -207,13 +330,18 @@ iterator frames*(kea: Kea): Frame =
 
     previousTime = currentTime
 
+    kea.mouse.beginFrame()
+    kea.keyboard.beginFrame()
+
     glfwPollEvents()
 
-    kea.input.update(kea.window)
+    kea.mouse.update(kea.window)
+    kea.keyboard.update(kea.window)
 
     yield Frame(
       delta: delta,
       fps: fps,
       time: time,
-      input: kea.input,
+      keyboard: kea.keyboard,
+      mouse: kea.mouse
     )
